refactor: consolidate StoreRedirectScreen logic into a single file
